refactor(dashboard): address review feedback on type-aware datasource resolution
Four review items from apache/superset#42539:

- The semanticViewDimensionsToColumns test now asserts type_generic for a
  temporal, string, and numeric dimension, using wire-shaped pyarrow type
  names (timestamp[us]/string/double) — a regression in
  mapSemanticTypeToGenericDataType fails the test instead of passing
  silently (control-run verified). SQL-style names like VARCHAR were never
  what /structure serialises, and the mapper rightly does not match them.

- The shared semantic-view structure helpers move to a layer-neutral module
  (src/utils/semanticViewStructure.ts); the filter-form utils re-export them
  for existing call sites, and the generic drill-info hook no longer imports
  from a dashboard modal utility.

- useDatasetDrillInfo's semantic branch builds a structurally valid Dataset
  with no cast: the fabricated id/datasource_type fields are dropped
  (consumers read only columns/metrics and derive verbose_map) and
  verbose_name is omitted rather than null.

- The FiltersConfigForm metric mapping documents its one genuine type gap
  instead of erasing the type: the /structure wire carries no metric uuid
  and this state's consumers never read it, so the cast narrows to
  Omit<Metric, 'uuid'> — every other field stays compiler-checked. (The
  blanket unknown-cast was hiding the uuid gap, not the verbose_name one.)

- getDatasourceTypeFromId renamed to getDatasourceTypeFromDatasourceId: it
  parses the composite `<id>__<type>` datasource string, not a plain id.
