+++ superset-frontend/src/dashboard/components/nativeFilters/FiltersConfigModal/FiltersConfigForm/utils.test.ts
@@ -419,21 +419,35 @@ test('fetchSemanticViewStructure defaults missing arrays to empty', async () => 
 
 test('semanticViewDimensionsToColumns maps dimension fields incl. temporal detection', () => {
   const { semanticViewDimensionsToColumns: toColumns } = require('./utils');
+  // Wire-shaped types: /structure serialises pyarrow types
+  // (timestamp[us], string, double), not SQL type names.
   const columns = toColumns([
-    { name: 'ordered_at', type: 'TIMESTAMP' },
-    { name: 'status', type: 'VARCHAR' },
+    { name: 'ordered_at', type: 'timestamp[us]' },
+    { name: 'status', type: 'string' },
+    { name: 'amount', type: 'double' },
   ]);
 
+  // type_generic is asserted explicitly: downstream filter-type logic (e.g.
+  // the Numerical range column filter) selects on it, so a regression in
+  // mapSemanticTypeToGenericDataType must fail here, not pass silently.
   expect(columns[0]).toMatchObject({
     column_name: 'ordered_at',
-    type: 'TIMESTAMP',
+    type: 'timestamp[us]',
     is_dttm: true,
     filterable: true,
+    type_generic: GenericDataType.Temporal,
   });
   expect(columns[1]).toMatchObject({
     column_name: 'status',
     is_dttm: false,
     filterable: true,
+    type_generic: GenericDataType.String,
+  });
+  expect(columns[2]).toMatchObject({
+    column_name: 'amount',
+    is_dttm: false,
+    filterable: true,
+    type_generic: GenericDataType.Numeric,
   });
 });
