+++ superset-frontend/src/dashboard/components/nativeFilters/FiltersConfigModal/FiltersConfigForm/utils.ts
@@ -23,7 +23,6 @@ import { CustomControlItem, Dataset } from '@superset-ui/chart-controls';
 import { Column, DatasourceType, ensureIsArray } from '@superset-ui/core';
 import { GenericDataType } from '@apache-superset/core/common';
 import { DatasourcesState, ChartsState } from 'src/dashboard/types';
-import { cachedSupersetGet } from 'src/utils/cachedSupersetGet';
 import { FILTER_SUPPORTED_TYPES } from './constants';
 
 const FILTERS_FIELD_NAME = 'filters';
@@ -102,94 +101,15 @@ export const doesColumnMatchFilterType = (filterType: string, column: Column) =>
     filterType as keyof typeof FILTER_SUPPORTED_TYPES
   ]?.includes(column.type_generic);
 
-export const mapSemanticTypeToGenericDataType = (
-  semanticType?: string | null,
-): GenericDataType | undefined => {
-  if (!semanticType) {
-    return undefined;
-  }
-
-  const normalized = semanticType.toLowerCase();
-
-  if (
-    /^(struct|list|map|array|fixed_size_list|large_list|union|dictionary)\b/.test(
-      normalized,
-    )
-  ) {
-    return undefined;
-  }
-
-  if (normalized.includes('bool')) {
-    return GenericDataType.Boolean;
-  }
-
-  if (/(date|time|timestamp|datetime)/.test(normalized)) {
-    return GenericDataType.Temporal;
-  }
-
-  if (
-    /(\b(u?int\d*)\b|\bfloat\d*\b|\bdouble\b|\bdecimal\d*\b|\bnumber\b)/.test(
-      normalized,
-    )
-  ) {
-    return GenericDataType.Numeric;
-  }
-
-  if (
-    /(\bstr(ing)?\b|\butf8\b|\blarge_string\b|\bbinary\b|\bjson\b|\buuid\b)/.test(
-      normalized,
-    )
-  ) {
-    return GenericDataType.String;
-  }
-
-  return undefined;
-};
-
-/** The slice of `GET /api/v1/semantic_view/<id>/structure` consumers rely on. */
-export interface SemanticViewStructure {
-  name?: string;
-  dimensions: { name: string; type: string }[];
-  metrics: { name: string; definition: string }[];
-}
-
-/**
- * Fetch a semantic view's structure — the single, shared entry point for
- * every type-aware datasource consumer (ColumnSelect, FiltersConfigForm,
- * display controls, drill metadata). Semantic views and regular datasets
- * have independent numeric-id sequences, so callers must route here (and
- * never to `/api/v1/dataset/<id>`) whenever the binding's datasourceType
- * is SemanticView.
- */
-export const fetchSemanticViewStructure = async (
-  semanticViewId: number | string,
-): Promise<SemanticViewStructure> => {
-  const response = await cachedSupersetGet({
-    endpoint: `/api/v1/semantic_view/${semanticViewId}/structure`,
-  });
-  const { name, dimensions = [], metrics = [] } = response.json?.result ?? {};
-  return { name, dimensions, metrics };
-};
-
-/**
- * Map structure dimensions onto the `Column` shape the filter/control
- * machinery expects. Mapping semantics are shared by every consumer:
- * temporal detection via `mapSemanticTypeToGenericDataType`, and
- * `filterable: true` (semantic-view dimensions are always selectable).
- */
-export const semanticViewDimensionsToColumns = (
-  dimensions: SemanticViewStructure['dimensions'],
-): Column[] =>
-  dimensions.map(dim => {
-    const mappedType = mapSemanticTypeToGenericDataType(dim.type);
-    return {
-      column_name: dim.name,
-      type: dim.type,
-      is_dttm: mappedType === GenericDataType.Temporal,
-      type_generic: mappedType,
-      filterable: true,
-    };
-  });
+// Shared semantic-view structure helpers live in a layer-neutral module so
+// non-dashboard consumers (e.g. the dataset drill-info hook) need not import
+// from this filter-form utility. Re-exported here for existing call sites.
+export {
+  mapSemanticTypeToGenericDataType,
+  fetchSemanticViewStructure,
+  semanticViewDimensionsToColumns,
+} from 'src/utils/semanticViewStructure';
+export type { SemanticViewStructure } from 'src/utils/semanticViewStructure';
 
 // Validates that a filter default value is present when the default value option is enabled.
 // For range filters, at least one of the two values must be non-null.
